+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - Add ability to search for coloreless color identity with `ci:c` and `ci:colorless`
   - Add error messages for errors in query parsing
   - Add ability to search by number of cards in a combo
+  - Support `cards` as alias for `card`
+  - Support `-cards` as alias for `-card`
 
 _Breaking Changes_